split off build related code into separate module
